+++ NEWS.md
@@ -6,6 +6,7 @@
   * NEWS file was converted to .md file.
   * Users can now provide a title for the plot.
   * Users can now provide titles for the primary and secondary y-axis.
+  * Changed logical checks `x == TRUE` and `x == FALSE` to `isTRUE(x)` and to `isFALSE(x)`.
 
 1.2.0
 -------------